# internal/features/tasks/transport/http/patch_task.go
package http

import (
	"fmt"
	"net/http"
	"unicode/utf8"

	"github.com/sparxfort1ano/go-todoapp/internal/core/domain"
	"github.com/sparxfort1ano/go-todoapp/internal/core/logger"
	"github.com/sparxfort1ano/go-todoapp/internal/core/transport/http/request"
	"github.com/sparxfort1ano/go-todoapp/internal/core/transport/http/response"
	"github.com/sparxfort1ano/go-todoapp/internal/core/transport/http/types"
)

// PatchTaskRequest represents the incoming JSON body for a partial task update (DTO).
type PatchTaskRequest struct {
	Title       types.Nullable[string] `json:"title" swaggertype:"string" example:"Погулять с собакой"`
	Description types.Nullable[string] `json:"description" swaggertype:"string" example:"null"`
	Completed   types.Nullable[bool]   `json:"completed" swaggertype:"boolean"`
}

// Validate performs early HTTP-level validation on the incoming payload.
// It ensures that string lengths are correct and required fields are not getting deleted
// before passing the data down to the domain layer.
func (r *PatchTaskRequest) Validate() error {
	if r.Title.Set {
		if r.Title.Value == nil {
			return fmt.Errorf("'Title' can't be NULL")
		}

		titleLen := utf8.RuneCountInString(*r.Title.Value)
		if titleLen < 1 || titleLen > 100 {
			return fmt.Errorf("'Title' must be between 1 and 100 symbols")
		}
	}

	if r.Description.Set {
		if r.Description.Value != nil {
			descriptionLen := utf8.RuneCountInString(*r.Description.Value)
			if descriptionLen < 1 || descriptionLen > 1000 {
				return fmt.Errorf("'Description' must be between 1 and 1000 symbols")
			}
		}
	}

	if r.Completed.Set {
		if r.Completed.Value == nil {
			return fmt.Errorf("'Completed' can't be NULL")
		}
	}

	return nil
}

// PatchTaskResponse represents the outgoing JSON body after a partial task update (JSON).
type PatchTaskResponse TaskDTOResponse

// PatchTask processes the HTTP request to partially update an existing task
// with the given id, validating the incoming JSON body.
//
// @Summary		Обновить задачу
// @Description Обновляет информацию об уже существующей в системе задаче.
// @Description ### Логика обновления полей (Three-state logic):
// @Description 1. **Поле не передано**: `description` игнорируется, значение в БД не меняется.
// @Description 2. **Явно передано значение**: `"description": "Утром в 6:30 выйти на прогулку с Бобиком"` - устанавливает новое описание для задачи.
// @Description 3. **Передан null**: `"description": null` - очищает поле в БД (set to NULL).
// @Description Ограничения: `title` и `completed` не могут быть выставлены как null.
// @Tags 		tasks
// @Accept 		json
// @Produce 	json
// @Param 		id 		path int 			  true  "ID изменяемой задачи"
// @Param		request body PatchTaskRequest true 	"PatchTask тело запроса"
// @Success 	200 {object} PatchTaskResponse 		"Успешно измененная задача"
// @Failure 	400 {object} response.ErrorResponse "Bad request"
// @Failure 	404 {object} response.ErrorResponse "Task not found"
// @Failure		409 {object} response.ErrorResponse "Conflict"
// @Failure 	500 {object} response.ErrorResponse "Internal server error"
// @Router 		/tasks/{id}	 [patch]
func (h *TasksHTTPHandler) PatchTask(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := logger.FromContext(ctx)
	responseHandler := response.NewHTTPResponseHandler(log, w)

	taskID, err := request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get taskID path value",
		)
		return
	}

	var req PatchTaskRequest
	if err := request.DecodeAndValidateRequest(r, &req); err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to decode and validate HTTP request",
		)
		return
	}

	taskPatch := taskPatchFromRequest(req)
	taskDomain, err := h.tasksService.PatchTask(ctx, taskID, taskPatch)
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to patch task",
		)
		return
	}

	response := PatchTaskResponse(taskDTOFromDomain(taskDomain))
	responseHandler.JSONResponse(response, http.StatusOK)
}

func taskPatchFromRequest(req PatchTaskRequest) domain.TaskPatch {
	return domain.NewTaskPatch(
		req.Title.ToDomain(),
		req.Description.ToDomain(),
		req.Completed.ToDomain(),
	)
}